+++ deltachat-ios/Controller/Settings/AdvancedViewController.swift
@@ -13,7 +13,6 @@ internal final class AdvancedViewController: UITableViewController {
 
     private enum CellTags: Int {
         case defaultTagValue = 0
-        case showEmails
         case viewLog
         case transportSettings
         case proxySettings
@@ -26,15 +25,6 @@ internal final class AdvancedViewController: UITableViewController {
     private let externalPathDescr = "File Sharing/Delta Chat"
 
     // MARK: - cells
-    private lazy var showEmailsCell: UITableViewCell = {
-        let cell = UITableViewCell(style: .value1, reuseIdentifier: nil)
-        cell.tag = CellTags.showEmails.rawValue
-        cell.textLabel?.text = String.localized("pref_show_emails")
-        cell.accessoryType = .disclosureIndicator
-        cell.detailTextLabel?.text = EmailOptionsViewController.getValString(val: dcContext.showEmails)
-        return cell
-    }()
-
     private lazy var transportSettingsCell: UITableViewCell = {
         let cell = UITableViewCell(style: .default, reuseIdentifier: nil)
         cell.textLabel?.text = String.localized("transports")
@@ -68,72 +58,6 @@ internal final class AdvancedViewController: UITableViewController {
                     alert.addAction(UIAlertAction(title: String.localized("cancel"), style: .cancel, handler: { [weak self] _ in
                         cell.uiSwitch.setOn(true, animated: true)
                     }))
-                    self.navigationController?.present(alert, animated: true, completion: nil)
-                }
-        })
-    }()
-
-    lazy var mvboxMoveCell: SwitchCell = {
-        return SwitchCell(
-            textLabel: String.localized("pref_auto_folder_moves"),
-            on: dcContext.getConfigBool("mvbox_move"),
-            action: { cell in
-                self.dcContext.setConfigBool("mvbox_move", cell.isOn)
-        })
-    }()
-
-    lazy var onlyFetchMvboxCell: SwitchCell = {
-        return SwitchCell(
-            textLabel: String.localized("pref_only_fetch_mvbox_title"),
-            on: dcContext.getConfigBool("only_fetch_mvbox"),
-            action: { cell in
-                if cell.isOn {
-                    let alert = UIAlertController(title: String.localized("pref_only_fetch_mvbox_title"),
-                        message: String.localized("pref_imap_folder_warn_disable_defaults"),
-                        preferredStyle: .alert)
-                    alert.addAction(UIAlertAction(title: String.localized("perm_continue"), style: .destructive, handler: { [weak self] _ in
-                        self?.dcContext.setConfigBool("only_fetch_mvbox", true)
-                    }))
-                    alert.addAction(UIAlertAction(title: String.localized("cancel"), style: .cancel, handler: { [weak self] _ in
-                        cell.uiSwitch.setOn(false, animated: true)
-                    }))
-                    self.navigationController?.present(alert, animated: true, completion: nil)
-                } else {
-                    self.dcContext.setConfigBool("only_fetch_mvbox", false)
-                }
-        })
-    }()
-
-    lazy var broadcastListsCell: SwitchCell = {
-        return SwitchCell(
-            textLabel: String.localized("channels"),
-            on: UserDefaults.standard.bool(forKey: "broadcast_lists"),
-            action: { cell in
-                UserDefaults.standard.set(cell.isOn, forKey: "broadcast_lists")
-                if cell.isOn {
-                    let alert = UIAlertController(title: "Thanks for trying out experimental 🧪 \"Channels\"!",
-                        message: "You can now create new \"Channels\" from the \"New Chat\" dialog\n\n"
-                               + "If you want to quit the experimental feature, you can disable it at \"Settings / Advanced\".",
-                        preferredStyle: .alert)
-                    alert.addAction(UIAlertAction(title: String.localized("ok"), style: .default, handler: nil))
-                    self.navigationController?.present(alert, animated: true, completion: nil)
-                }
-        })
-    }()
-
-    lazy var callsCell: SwitchCell = {
-        return SwitchCell(
-            textLabel: "Debug Calls",
-            on: UserDefaults.standard.bool(forKey: "pref_calls_enabled"),
-            action: { cell in
-                UserDefaults.standard.set(cell.isOn, forKey: "pref_calls_enabled")
-                if cell.isOn {
-                    let alert = UIAlertController(title: "Thanks for helping to debug 🧪 \"Calls\"!",
-                        message: "You can now debug calls using the phone-icon in one-to-one-chats\n\n"
-                               + "The experiment is about making decentralised calls work and reliable at all, not about options or UI. "
-                               + "We're happy about focused feedback at support.delta.chat",
-                        preferredStyle: .alert)
-                    alert.addAction(UIAlertAction(title: String.localized("ok"), style: .default, handler: nil))
                     self.navigationController?.present(alert, animated: true, completion: nil)
                 }
         })
@@ -185,17 +109,9 @@ internal final class AdvancedViewController: UITableViewController {
         let experimentalSection = SectionConfigs(
             headerTitle: String.localized("pref_experimental_features"),
             footerTitle: String.localized("pref_experimental_features_explain"),
-            cells: [broadcastListsCell, callsCell, locationStreamingCell, appPickerCell])
+            cells: [locationStreamingCell, appPickerCell])
 
-        if dcContext.isChatmail {
-            return [viewLogSection, serverSection, experimentalSection]
-        } else {
-            let legacySection = SectionConfigs(
-                headerTitle: "Legacy Options",
-                footerTitle: nil,
-                cells: [showEmailsCell, mvboxMoveCell, onlyFetchMvboxCell])
-            return [viewLogSection, serverSection, experimentalSection, legacySection]
-        }
+        return [viewLogSection, serverSection, experimentalSection]
     }()
 
     init(dcAccounts: DcAccounts) {
@@ -241,8 +157,6 @@ internal final class AdvancedViewController: UITableViewController {
         tableView.deselectRow(at: indexPath, animated: false)
 
         switch cellTag {
-        case .showEmails: showClassicMailController()
-
         case .viewLog: showLogViewController()
 
         case .transportSettings:
@@ -279,11 +193,6 @@ internal final class AdvancedViewController: UITableViewController {
 
     private func showLogViewController() {
         let controller = LogViewController(dcContext: dcContext)
-        navigationController?.pushViewController(controller, animated: true)
-    }
-
-    private func showClassicMailController() {
-        let controller = EmailOptionsViewController(dcContext: dcContext)
         navigationController?.pushViewController(controller, animated: true)
     }
 
@@ -325,7 +234,6 @@ internal final class AdvancedViewController: UITableViewController {
 
     // MARK: - updates
     private func updateCells() {
-        showEmailsCell.detailTextLabel?.text = EmailOptionsViewController.getValString(val: dcContext.showEmails)
         proxySettingsCell.detailTextLabel?.text = dcContext.isProxyEnabled ? String.localized("on") : nil
     }
 }