+++ deltachat-ios/Controller/Settings/ChatsAndMediaViewController.swift
@@ -15,7 +15,6 @@ internal final class ChatsAndMediaViewController: UITableViewController {
         case receiptConfirmation
         case exportBackup
         case autodelDevice
-        case autodelServer
         case mediaQuality
         case downloadOnDemand
     }
@@ -34,31 +33,12 @@ internal final class ChatsAndMediaViewController: UITableViewController {
         return cell
     }()
 
-    func autodelSummary() -> String {
-        let delDeviceAfter = dcContext.getConfigInt("delete_device_after")
-        let delServerAfter = dcContext.isChatmail ? 0 : dcContext.getConfigInt("delete_server_after")
-        if delDeviceAfter==0 && delServerAfter==0 {
-            return String.localized("never")
-        } else {
-            return String.localized("on")
-        }
-    }
-
     private lazy var autodelDeviceCell: UITableViewCell = {
         let cell = UITableViewCell(style: .subtitle, reuseIdentifier: nil)
         cell.tag = CellTags.autodelDevice.rawValue
         cell.textLabel?.text = String.localized("autodel_device_title")
         cell.accessoryType = .disclosureIndicator
-        cell.detailTextLabel?.text = AutodelOptionsViewController.getSummary(dcContext, fromServer: false)
-        return cell
-    }()
-
-    private lazy var autodelServerCell: UITableViewCell = {
-        let cell = UITableViewCell(style: .subtitle, reuseIdentifier: nil)
-        cell.tag = CellTags.autodelServer.rawValue
-        cell.textLabel?.text = String.localized("autodel_server_title")
-        cell.accessoryType = .disclosureIndicator
-        cell.detailTextLabel?.text = AutodelOptionsViewController.getSummary(dcContext, fromServer: true)
+        cell.detailTextLabel?.text = AutodelOptionsViewController.getSummary(dcContext)
         return cell
     }()
 
@@ -112,7 +92,7 @@ internal final class ChatsAndMediaViewController: UITableViewController {
         let autoDelSection = SectionConfigs(
             headerTitle: String.localized("delete_old_messages"),
             footerTitle: nil,
-            cells: dcContext.isChatmail ? [autodelDeviceCell] : [autodelDeviceCell, autodelServerCell]
+            cells: [autodelDeviceCell]
         )
         let exportBackupSection = SectionConfigs(
             headerTitle: nil,
@@ -171,11 +151,7 @@ internal final class ChatsAndMediaViewController: UITableViewController {
         case .receiptConfirmation: break
 
         case .autodelDevice:
-            let controller = AutodelOptionsViewController(dcContext: dcContext, fromServer: false)
-            navigationController?.pushViewController(controller, animated: true)
-
-        case .autodelServer:
-            let controller = AutodelOptionsViewController(dcContext: dcContext, fromServer: true)
+            let controller = AutodelOptionsViewController(dcContext: dcContext)
             navigationController?.pushViewController(controller, animated: true)
 
         case .exportBackup:
@@ -190,11 +166,7 @@ internal final class ChatsAndMediaViewController: UITableViewController {
     }
 
     override func tableView(_ tableView: UITableView, titleForFooterInSection section: Int) -> String? {
-        if sections[section].cells.last == autodelServerCell && dcContext.getConfigInt("delete_server_after") != 0 {
-            return String.localized("autodel_server_enabled_hint")
-        } else {
-            return sections[section].footerTitle
-        }
+        return sections[section].footerTitle
     }
 
     // MARK: - actions
@@ -248,8 +220,7 @@ internal final class ChatsAndMediaViewController: UITableViewController {
         mediaQualityCell.detailTextLabel?.text = MediaQualityViewController.getValString(val: dcContext.getConfigInt("media_quality"))
         downloadOnDemandCell.detailTextLabel?.text = DownloadOnDemandViewController.getValString(
             val: dcContext.getConfigInt("download_limit"))
-        autodelDeviceCell.detailTextLabel?.text = AutodelOptionsViewController.getSummary(dcContext, fromServer: false)
-        autodelServerCell.detailTextLabel?.text = AutodelOptionsViewController.getSummary(dcContext, fromServer: true)
+        autodelDeviceCell.detailTextLabel?.text = AutodelOptionsViewController.getSummary(dcContext)
     }
 
     // MARK: - coordinator